rn-801: Add data grid to viz builder
